+++ books_backend/src/main/java/com/psuti/books/controller/UserAddressController.java
@@ -6,6 +6,7 @@ import com.psuti.books.model.OfferList;
 import com.psuti.books.model.UserAddress;
 import com.psuti.books.security.UserPrincipal;
 import com.psuti.books.service.UserAddressService;
+import com.psuti.books.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,19 +21,23 @@ import java.util.List;
 public class UserAddressController {
 
     private final UserAddressService userAddressService;
+    private final UserService userService;
 
     @GetMapping
     public ResponseEntity<List<UserAddress>> getAddressesByUserPrincipal(@AuthenticationPrincipal UserPrincipal principal) {
+        if (!userService.checkEnabledPrincipal(principal)) return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);//Проверка на бан
         return new ResponseEntity<>(userAddressService.getByUserPrincipal(principal), HttpStatus.OK);
     }
 
     @PostMapping
     public ResponseEntity<UserAddress> createUserAddress(@RequestBody UserAddressDTO dto, @AuthenticationPrincipal UserPrincipal principal) {
+        if (!userService.checkEnabledPrincipal(principal)) return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);//Проверка на бан
         return new ResponseEntity<>(userAddressService.create(dto, principal), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public HttpStatus deleteUserAddress(@PathVariable Long id, @AuthenticationPrincipal UserPrincipal principal) {
+        if (!userService.checkEnabledPrincipal(principal)) return HttpStatus.FORBIDDEN;//Проверка на бан
         userAddressService.delete(id, principal);
         return HttpStatus.OK;
     }
